projeto com as cruds finalizadas

+++ src/medias/medias.service.ts
@@ -1,4 +1,4 @@
-import { ConflictException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, ForbiddenException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateMediaDto } from './dto/create-media.dto';
 import { UpdateMediaDto } from './dto/update-media.dto';
 import { MediasRepository } from './medias.repository';
@@ -15,25 +15,14 @@ export class MediasService {
   };
 
   async findAll() {
-    const medias = await this.repository.findMedias();
-    return medias.map((media) => ({
-      id: media.id,
-      title: media.title,
-      username: `https://www.${media.title.toLowerCase()}.com/${media.username}`
-    }));
+    return await this.repository.findMedias();
   };
 
   async findOne(id: number) {
     const media = await this.repository.findMediaById(id);
     if(!media) throw new NotFoundException("Id not found!");
 
-    const data = [{
-      id: media.id,
-      title: media.title,
-      username: `https://www.${media.title.toLowerCase()}.com/${media.username}`
-    }];
-
-    return data;
+    return media;
   };
 
   async update(id: number, body: UpdateMediaDto) {
@@ -50,6 +39,9 @@ export class MediasService {
     const media = await this.repository.findMediaById(id);
     if(!media) throw new NotFoundException("Id not found!");
 
+    const publication = await this.repository.findPublicationsById(id);
+    if(publication) throw new ForbiddenException();
+
     return await this.repository.deleteMedia(id);
-  }
-}
+  };
+};
